Check StandardID and Len in VENDOR_DEFINED_*
Fix #3039.

+++ include/internal/libspdm_common_lib.h
@@ -1938,11 +1938,11 @@ uint32_t libspdm_mask_base_asym_algo(libspdm_context_t *spdm_context, uint32_t b
  * Check if the combination of SVH ID and VendorIDLen are legal.
  *
  * @param  id             Registry or standards body identifier (SPDM_REGISTRY_ID_*).
- * @param  vendor_id_len  Length, in bytes, of the VendorID field.
- *
+ *                        Its size is two bytes due to the vendor-defined messages.
+ * @param  vendor_id_len  Length, in bytes, of the VendorID field. *
  * @retval true  The ID and VendorIDLen are legal.
  * @retval false The ID and VendorIDLen are illegal.
  */
-bool libspdm_validate_svh_vendor_id_len(uint8_t id, uint8_t vendor_id_len);
+bool libspdm_validate_svh_vendor_id_len(uint16_t id, uint8_t vendor_id_len);
 
 #endif /* SPDM_COMMON_LIB_INTERNAL_H */
